UPD: conformity_scores module migration
