+++ mapie/conformity_scores/__init__.py
@@ -1,0 +1,9 @@
+from .conformity_scores import ConformityScore
+from .residual_conformity_scores import (AbsoluteConformityScore,
+                                         GammaConformityScore)
+
+__all__ = [
+    "ConformityScore",
+    "AbsoluteConformityScore",
+    "GammaConformityScore"
+]
